+++ src/pages/chatting/components/styles/chattingRoom.module.css
@@ -48,3 +48,14 @@
   font-size: 11px;
   align-items: flex-end;
 }
+
+.messages_container {
+  display: flex;
+  flex-direction: column;
+}
+
+.dateTitle {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
