+++ src/pages/home/Home.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import styles from './Home.module.css';
+import styles from './styles/Home.module.css';
 
 const Home = () => {
   return (
