--- src/pages/match/components/badgeList.module.css
@@ -1,10 +1,0 @@
-.container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-items: center;
-  gap: 0.3rem;
-  padding: 0.5rem;
-  margin: 0px;
-
-  cursor: default;
-}
