+++ src/pages/match/message/components/messagePreview.module.css
@@ -61,7 +61,6 @@
 .date {
   font-size: 11px;
   text-align: center;
-
   color: gray;
 }
